% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvec.R
\name{applyFun}
\alias{applyFun}
\title{Apply function to arguments that can be vectors and lists}
\usage{
applyFun(fun, ..., atoms = c(""))
}
\arguments{
\item{fun}{Function name}

\item{...}{Argumetns that can be vectors and lists}

\item{atoms}{Vector of class names that cannot be broken apart by index referencing}
}
\value{
Function's result that can be vector or list
}
\description{
Apply function to arguments that can be vectors and lists
}
