% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvec.R
\name{getAttr}
\alias{getAttr}
\title{Return value of object's attribute}
\usage{
getAttr(obj, attrName)
}
\arguments{
\item{obj}{Object or list of objects}

\item{attrName}{Attribute name as character}
}
\value{
Attribute of object
}
\description{
Return value of object's attribute
}
