% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvec.R
\name{getLens}
\alias{getLens}
\title{Return vector of lengths of list elements}
\usage{
getLens(l)
}
\arguments{
\item{l}{List}
}
\value{
Vector of lengths
}
\description{
Return vector of lengths of list elements
}
