% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvec.R
\name{isVector}
\alias{isVector}
\title{Return TRUE if variable is vector or Date}
\usage{
isVector(v)
}
\arguments{
\item{v}{Any variable}
}
\value{
TRUE if variable is vector or Date
}
\description{
Return TRUE if variable is vector or Date
}
