% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvec.R
\name{applyFun}
\alias{applyFun}
\title{Apply function to arguments that can be vectors}
\usage{
applyFun(fun, ...)
}
\arguments{
\item{fun}{Function name}

\item{...}{Argumetns that can be vectors}
}
\value{
Function's result or results as vector
}
\description{
Apply function to arguments that can be vectors
}
