% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvec.R
\name{getElem}
\alias{getElem}
\title{Return element "i" of "a" assuming that "a" has infinite tail filled with its last element}
\usage{
getElem(a, i)
}
\arguments{
\item{a}{Any vector}

\item{i}{Index of value}
}
\value{
Value with index i
}
\description{
Return element "i" of "a" assuming that "a" has infinite tail filled with its last element
}
