% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvec.R
\name{getLen}
\alias{getLen}
\title{Check that all params have length = 1 or same length > 1}
\usage{
getLen(...)
}
\arguments{
\item{...}{Any number of vectors}
}
\value{
Maximum length of params
}
\description{
Check that all params have length = 1 or same length > 1
}
