% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvec.R
\name{getLens}
\alias{getLens}
\title{Return vector of lengths of list elements}
\usage{
getLens(l, atoms = c(""))
}
\arguments{
\item{l}{List}

\item{atoms}{Vector of class names that lengths are assumed to be 1}
}
\value{
Vector of lengths
}
\description{
Return vector of lengths of list elements
}
